feat(feed): Awaiting a feed of an non nullable type, now return T? instead of Option<T>
BREAKING: Awaiting a feed of an non nullable type, now return T? instead of Option<T>, there is no longer a way to await a feed of a nullable T, you have instead to use the new .Option().
None and Undefined are mapped to null, a new .Option() is now available to easily get the option if needed.