feat(TokenCacheExtensions): added optional parameters to the SaveTokensAsync method that will allow to use it with own Keys chore: Added Extensions for the other standarized Keys alongside with the relevant const string representations chore: Added to each of the seperate existing methods for e.g. AccessTokenAsync the key name as optional parameter
By allowing the key as optional parameter and those new extensions, we will gain more flexibility for the user of it.
This will be especially usefull for cases like oAuth where we do not have a id_token then e.g. a user_id that would be important to store instead of that and would represent in this scenario the equivalent to the id_token of jwt.