chore: change the accessor of the WebConfiguration to public
By this, it not longer required for the User of this extension, to create a equal own definition just to get those defined values loaded from Appsettings. This is needed for example if those values are also needed in the app not just in the extension which is using this in first place. also this allowes to use intellisense and decompilation view to see what properties we can provide for confiuration without having to dig up the source code