chore: preserve properties on IListFeed<T>
It's our favorite exercise, making uno.chefs run under NativeAOT!

	dotnet publish -c Release -r osx-x64 -f net10.0-desktop -p:TargetFrameworkOverride=net10.0-desktop -bl \
	  Chefs/Chefs.csproj \
	  -p:SelfContained=true -p:PublishAot=true -p:IsAotCompatible=true -p:UseSkiaRendering=true \
	  -p:IlcGenerateMapFile=true -p:IlcGenerateMstatFile=true -p:IlcGenerateDgmlFile=true -p:IlcGenerateMetadataLog=true \
	  -p:EmitCompilerGeneratedFiles=true -p:CompilerGeneratedFilesOutputPath=`pwd`/_gen
	Chefs/bin/Release/net10.0-desktop/osx-x64/publish/Chefs

Login to the app, click on a recipe, and no Ingredients are shown.

The app log contains messages such as:

	fail: Uno.UI.DataBinding.BindingPropertyHelper[0]
	      The [Name] property getter does not exist on type [Chefs.Business.Models.Ingredient]

The problem is that the properties on `Ingredient` are not being
preserved.

The question: how "best" to preserve them?

As @jonpryor is the "automagic" sort, the "obvious" (and possibly
incorrect) answer is "something *outside* of uno.chefs should ensure
that the properties are preserved".

But where is `Ingredient` mentioned within uno.chefs?  What types
can be (ab)used to preserve properties?

Turns out, not too many places:

	% git grep '\<Ingredient\>'
	Chefs/Business/Models/Ingredient.cs:public record Ingredient
	Chefs/Business/Models/Ingredient.cs:    public Ingredient(IngredientData ingredientData)
	Chefs/Business/Services/Recipes/IRecipeService.cs:      ValueTask<IImmutableList<Ingredient>> GetIngredients(Guid recipeId, CancellationToken ct);
	Chefs/Business/Services/Recipes/RecipeService.cs:       public async ValueTask<IImmutableList<Ingredient>> GetIngredients(Guid recipeId, CancellationToken ct)
	Chefs/Business/Services/Recipes/RecipeService.cs:               return ingredientsData?.Select(x => new Ingredient(x)).ToImmutableList() ?? ImmutableList<Ingredient>.Empty;
	Chefs/Presentation/RecipeDetailsModel.cs:       public IListFeed<Ingredient> Ingredients => ListFeed.Async(async ct => await _recipeService.GetIngredients(Recipe.Id, ct));

The `Ingredient` type definition won't help, `IImmutableList<T>`
would be a *massive* stretch (plus we don't own it), which brings us
to `IListFeed<T>`: that *is* a type that we control!

Update `IListFeed<T>` to preserve properties on `T`:

	partial interface IListFeed<
	  [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicProperties)]
	> {}

Perhaps unsurprisingly, that change impacts *everything*.
(The patch size escalated quickly.)

Integrating this change back into uno.chefs, and now I'm able to see
Ingredients, without other changes to uno.chefs!  (At least it worked.)

But!  We next see this message in the app log:

	fail: Uno.UI.DataBinding.BindingPropertyHelper[0]
	      The [Count] property getter does not exist on type [Uno.Extensions.Reactive.Bindings.BindableListFeed`1[Chefs.Business.Models.Ingredient]]

Similar question as before: how do we preserve the properties on
`BindableListFeed<T>`?

Raising the question: where is `BindableListFeed<T>` coming from?

This was *quite* interesting, as `BindableListFeed<T>` doesn't appear
*anywhere* in uno.chefs!

	% git grep BindableListFeed
	# no source code matches
	% grep -ril BindableListFeed _gen
	# no matches

Add some printf logging to the `BindableListFeed<T>` constructors to
see how it's constructed, and:

	   at Uno.Extensions.Reactive.Bindings.BindableListFeed`1..ctor(String, IListState`1) + 0x91
	   at Uno.Extensions.Reactive.Bindings.BindableFactory.CreateList[T](String, IListState`1) + 0x2d
	   at Chefs.Presentation.RecipeDetailsViewModel..ctor(RecipeDetailsModel) + 0x333
	   at Chefs.Presentation.RecipeDetailsViewModel..ctor(Recipe recipe, INavigator navigator, IRecipeService recipeService, IUserService userService, IMessenger messenger, IShareService shareService) + 0xa0

`BindableListFeed<Ingredient>` is constructed via a
`BindableFactory.CreateList(…)` invocation within the generated
`RecipeDetailsViewModel` constructor.

Slight problem: there is no `BindableFactory.CreateList(…)` invocation
within the `RecipeDetailsViewModel` constructor!

	public partial class RecipeDetailsViewModel {
	  protected RecipeDetailsViewModel(global::Chefs.Presentation.RecipeDetailsModel model)
	  {
	    // …
	    if (Ingredients is null)
	    {
	      var ingredientsSource = (global::Uno.Extensions.Reactive.IListFeed<global::Chefs.Business.Models.Ingredient>) model.Ingredients ?? throw new NullReferenceException("The list feed property 'Ingredients' is null. Public feeds properties must be initialized in the constructor.");
	      var ingredientsSourceListState = ctx.GetOrCreateListState(ingredientsSource);
	      Ingredients = global::Uno.Extensions.Reactive.Bindings.BindableHelper.CreateBindableList(nameof(Ingredients), ingredientsSourceListState);
	    }
	    // …
	  }
	}

However, `BindableHelper.CreateBindableList(…)` *does* call
`BindableFactory.CreateList(…)`, so this is presumably a case of
method inlining by NativeAOT.

With that in mind, @jonpryor can think of three ways to cause the
`BindableListFeed<Ingredient>` properties to be preserved:

 1. Some form of change within FeedsGenerator, or
 2. Update `BindableFactory.CreateList<T>()` to use
    `[DynamicDependency]` to preserve the properties, or
 3. Update the `BindableListFeed<T>` constructors to have
    `[DynamicDependency]`.

The advantages to (3) is that it works, it's straightforward, and
annotation site is "close" to the site of intended effect.
(`BindableFactory.CreateList<T>()` doesn't obviously scream out
"I may require dynamic dependencies!"  The constructor doesn't either,
but at least it's on the type being impacted!)

(3) also gives us an interesting "hammer" with which to solve related
issues.  Issues such as:

	fail: Uno.UI.DataBinding.BindingPropertyHelper[0]
	      The [Count] property getter does not exist on type [Uno.Extensions.Reactive.Bindings.Collections._BindableCollection.Views.BasicView]

@jonpryor has no idea where `BasicView` is constructed from, but if
it's constructed, we *probably* need its properties, so: add
`[DynamicDependency]` to the `BasicView` constructor!  The message
is fixed.  (This is a fascinating hammer, ripe for abuse.)